Moves RouterFederatedStateProto to hadoop-rbf module. Adds more documentation.
